refactor(composables): fixed folder naming typo
