+++ packages/composables/src/composables/getters/index.ts
@@ -7,6 +7,7 @@ export { default as facetGetters } from './facetGetters';
 export { default as orderGetters } from './orderGetters';
 export { default as productGetters } from './productGetters';
 export { default as reviewGetters } from './reviewGetters';
+export { default as userAddressesGetters } from './userAddressesGetters';
 export { default as userBillingGetters } from './userBillingGetters';
 export { default as userGetters } from './userGetters';
 export { default as userShippingGetters } from './userShippingGetters';
